Project-OSdb/requirements.txt
# Core Dependencies

Pillow>=10.0.0
cryptography>=41.0.0
pyotp>=2.8.0
qrcode>=7.4.2

# Database


# API Integration
requests>=2.31.0
beautifulsoup4>=4.12.0
python-whois>=0.8.0
shodan>=1.29.0
virustotal-api>=1.1.11
tweepy>=4.14.0
python-linkedin>=1.0.1

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Scheduling
schedule>=1.2.0
email-validator>=2.0.0

# Security
scrypt>=0.8.24
bcrypt>=4.0.0
passlib>=1.7.4

# Networking
paramiko>=3.2.0
scapy>=2.5.0
dnspython>=2.4.0

# Additional
reportlab>=4.0.0
openpyxl>=3.1.0
pyzbar>=0.1.9
opencv-python>=4.8.0